# specs/002-intelligent-message-analysis/contracts/api-contract.yaml
openapi: 3.0.3
info:
  title: Intelligent Message Analysis API
  description: API endpoints for AI-powered message analysis, spam detection, and reply generation
  version: 1.0.0
  contact:
    name: MyAssistantPersonal Team
servers:
  - url: http://localhost:3002/api/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /settings:
    get:
      summary: Get user settings
      description: Retrieve current user's automation settings
      operationId: getUserSettings
      tags:
        - Settings
      responses:
        '200':
          description: Settings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettingsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update user settings
      description: Update automation settings for current user
      operationId: updateUserSettings
      tags:
        - Settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSettingsUpdateRequest'
      responses:
        '200':
          description: Settings updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettingsResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserSettingsResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/UserSettings'
    UserSettings:
      type: object
      required:
        - userId
        - autoDeleteSpamEnabled
        - spamThreshold
        - autoSendRepliesEnabled
        - responseConfidenceThreshold
        - autoReplyConditions
        - lastUpdated
      properties:
        userId:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
          example: '507f1f77bcf86cd799439011'
        autoDeleteSpamEnabled:
          type: boolean
          example: false
        spamThreshold:
          type: integer
          minimum: 0
          maximum: 100
          example: 80
        autoSendRepliesEnabled:
          type: boolean
          example: false
        responseConfidenceThreshold:
          type: integer
          minimum: 0
          maximum: 100
          example: 85
        autoReplyConditions:
          $ref: '#/components/schemas/AutoReplyConditions'
        lastUpdated:
          type: string
          format: date-time
          example: '2025-11-19T10:30:00Z'
    AutoReplyConditions:
      type: object
      required:
        - senderWhitelist
        - senderBlacklist
        - businessHoursOnly
        - maxRepliesPerDay
      properties:
        senderWhitelist:
          type: array
          items:
            type: string
            format: email
          example: ['important@example.com']
        senderBlacklist:
          type: array
          items:
            type: string
            format: email
          example: ['spam@example.com']
        businessHoursOnly:
          type: boolean
          example: false
        maxRepliesPerDay:
          type: integer
          minimum: 1
          maximum: 100
          example: 10
    UserSettingsUpdateRequest:
      type: object
      properties:
        autoDeleteSpamEnabled:
          type: boolean
        spamThreshold:
          type: integer
          minimum: 0
          maximum: 100
        autoSendRepliesEnabled:
          type: boolean
        responseConfidenceThreshold:
          type: integer
          minimum: 0
          maximum: 100
        autoReplyConditions:
          $ref: '#/components/schemas/AutoReplyConditions'
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - message
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Bad Request'
        message:
          type: string
          example: 'Invalid message ID format'
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Unauthorized
            message: Authentication token required
...
    BadRequestError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Bad Request
            message: spamThreshold must be between 0 and 100
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Internal Server Error
            message: An unexpected error occurred